Use dedicated settings file for tests
